出品者のみが削除できるよう編集

+++ app/controllers/products_controller.rb
@@ -36,8 +36,12 @@ class ProductsController < ApplicationController
   end
 
   def destroy
-    if @product.destroy
-      redirect_to root_path
+    if user_signed_in? && current_user.id == @product.id 
+      if @product.destroy
+        redirect_to root_path
+      end
+    else
+      render :show
     end
   end
